+++ cmd/yellow-silence/main.go
@@ -12,7 +12,7 @@ import (
 	"github.com/volodymyr/yellow-silence/internal/config"
 )
 
-const version = "0.1.0"
+var version = "0.1.0"
 
 func main() {
 	if err := run(); err != nil {